refactor: add action button to real time events
