+++ src/views/RealTimeEventsDataStream/ListView.vue
@@ -1,6 +1,6 @@
 <script setup>
   import EmptyResultsBlock from '@/templates/empty-results-block'
-  import ListTableBlock from '@/templates/list-table-block/no-header'
+  import ListTableBlock from '@/templates/list-table-block/action-column.vue'
   import Drawer from './Drawer'
   import PrimeButton from 'primevue/button'
   import { computed, ref } from 'vue'
@@ -135,6 +135,7 @@
     :editInDrawer="openDetailDrawer"
     @on-load-data="handleLoadData"
     emptyListMessage="No logs have been found for this period."
+    isTabs
   />
 
   <EmptyResultsBlock
